Repackage nodes
