+++ src/main/java/io/github/mattidragon/advancednetworking/graph/ModNodeTypes.java
@@ -1,8 +1,25 @@
 package io.github.mattidragon.advancednetworking.graph;
 
-import io.github.mattidragon.advancednetworking.graph.node.energy.*;
-import io.github.mattidragon.advancednetworking.graph.node.fluid.*;
-import io.github.mattidragon.advancednetworking.graph.node.item.*;
+import io.github.mattidragon.advancednetworking.graph.node.energy.filter.LimitEnergyNode;
+import io.github.mattidragon.advancednetworking.graph.node.energy.route.MergeEnergyNode;
+import io.github.mattidragon.advancednetworking.graph.node.energy.route.SplitEnergyNode;
+import io.github.mattidragon.advancednetworking.graph.node.energy.storage.EmptyEnergyStreamNode;
+import io.github.mattidragon.advancednetworking.graph.node.energy.storage.EnergySourceNode;
+import io.github.mattidragon.advancednetworking.graph.node.energy.storage.EnergyTargetNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.filter.FilterFluidNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.filter.LimitFluidNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.route.MergeFluidNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.route.SplitFluidNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.storage.EmptyFluidStreamNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.storage.FluidSourceNode;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.storage.FluidTargetNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.filter.FilterItemsNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.filter.LimitItemsNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.route.MergeItemsNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.route.SplitItemsNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.storage.EmptyItemStreamNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.storage.ItemSourceNode;
+import io.github.mattidragon.advancednetworking.graph.node.item.storage.ItemTargetNode;
 import io.github.mattidragon.advancednetworking.graph.node.redstone.ReadRedstoneNode;
 import io.github.mattidragon.advancednetworking.graph.node.redstone.SetRedstoneNode;
 import io.github.mattidragon.advancednetworking.graph.node.redstone.WriteRedstoneNode;
