+++ src/main/java/io/github/mattidragon/advancednetworking/graph/node/fluid/filter/LimitFluidNode.java
@@ -1,9 +1,10 @@
-package io.github.mattidragon.advancednetworking.graph.node.fluid;
+package io.github.mattidragon.advancednetworking.graph.node.fluid.filter;
 
 import com.mojang.datafixers.util.Either;
 import io.github.mattidragon.advancednetworking.client.screen.SliderConfigScreen;
 import io.github.mattidragon.advancednetworking.graph.ModDataTypes;
 import io.github.mattidragon.advancednetworking.graph.ModNodeTypes;
+import io.github.mattidragon.advancednetworking.graph.node.fluid.FluidTransformer;
 import io.github.mattidragon.nodeflow.graph.Connector;
 import io.github.mattidragon.nodeflow.graph.Graph;
 import io.github.mattidragon.nodeflow.graph.data.DataValue;
@@ -12,65 +13,59 @@ import io.github.mattidragon.nodeflow.ui.screen.EditorScreen;
 import io.github.mattidragon.nodeflow.ui.screen.NodeConfigScreen;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.transfer.v1.fluid.FluidConstants;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.text.Text;
 import net.minecraft.util.math.MathHelper;
 
 import java.util.List;
 
-public class SplitFluidNode extends Node {
-    private int count = 2;
+public class LimitFluidNode extends Node {
+    private int limit = (int) FluidConstants.BUCKET;
 
-    public SplitFluidNode(Graph graph) {
-        super(ModNodeTypes.SPLIT_FLUID, List.of(), graph);
+    public LimitFluidNode(Graph graph) {
+        super(ModNodeTypes.LIMIT_FLUID, List.of(), graph);
     }
 
     @Override
     public Connector<?>[] getOutputs() {
-        var connectors = new Connector[count];
-        for (int i = 0; i < connectors.length; i++) {
-            connectors[i] = ModDataTypes.FLUID_STREAM.makeOptionalOutput(String.valueOf(i), this);
-        }
-
-        return connectors;
+        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredOutput("out", this) };
     }
 
     @Override
     public Connector<?>[] getInputs() {
-        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredInput("items", this) };
+        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredInput("in", this) };
     }
 
     @Override
     protected Either<DataValue<?>[], Text> process(DataValue<?>[] inputs, ContextProvider context) {
-        var current = inputs[0].getAs(ModDataTypes.FLUID_STREAM);
-        var out = new DataValue<?>[count];
-
-        for (int i = 0; i < count - 1; i++) {
-            var split = current.split();
-            out[i] = ModDataTypes.FLUID_STREAM.makeValue(current);
-            current = split;
-        }
-        out[count - 1] = ModDataTypes.FLUID_STREAM.makeValue(current);
-
-        return Either.left(out);
+        var stream = inputs[0].getAs(ModDataTypes.FLUID_STREAM);
+        stream.transform(new FluidTransformer.Limit(limit));
+        return Either.left(new DataValue<?>[]{ ModDataTypes.FLUID_STREAM.makeValue(stream) });
     }
 
     @Override
     public void readNbt(NbtCompound data) {
         super.readNbt(data);
-        count = MathHelper.clamp(data.getInt("count"), 2, 8);
+        limit = MathHelper.clamp(data.getInt("limit"), 1, (int) FluidConstants.BUCKET);
     }
 
     @Override
     public void writeNbt(NbtCompound data) {
         super.writeNbt(data);
-        data.putInt("count", count);
+        data.putInt("limit", limit);
     }
 
     @Environment(EnvType.CLIENT)
     @Override
     public NodeConfigScreen createConfigScreen(EditorScreen parent) {
-        return new SliderConfigScreen(this, parent, value -> count = value, () -> count, Text.translatable("node.advanced_networking.streams"), 2, 8);
+        return new SliderConfigScreen(this,
+                parent,
+                value -> limit = (int) (Math.round(value / 1000.0) * 1000),
+                () -> limit,
+                Text.translatable("node.advanced_networking.limit"),
+                1,
+                (int) FluidConstants.BUCKET);
     }
 
     @Override
