+++ src/main/java/io/github/mattidragon/advancednetworking/graph/node/fluid/storage/FluidTargetNode.java
@@ -1,35 +1,32 @@
-package io.github.mattidragon.advancednetworking.graph.node.fluid;
+package io.github.mattidragon.advancednetworking.graph.node.fluid.storage;
 
 import com.mojang.datafixers.util.Either;
 import io.github.mattidragon.advancednetworking.graph.ModDataTypes;
 import io.github.mattidragon.advancednetworking.graph.ModNodeTypes;
 import io.github.mattidragon.advancednetworking.graph.NetworkControllerContext;
 import io.github.mattidragon.advancednetworking.graph.node.InterfaceNode;
-import io.github.mattidragon.advancednetworking.graph.path.PathBundle;
 import io.github.mattidragon.nodeflow.graph.Connector;
 import io.github.mattidragon.nodeflow.graph.Graph;
 import io.github.mattidragon.nodeflow.graph.context.ContextType;
 import io.github.mattidragon.nodeflow.graph.data.DataValue;
 import net.fabricmc.fabric.api.transfer.v1.fluid.FluidStorage;
-import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
-import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
 import net.minecraft.text.Text;
 
 import java.util.List;
 
-public class FluidSourceNode extends InterfaceNode {
-    public FluidSourceNode(Graph graph) {
-        super(ModNodeTypes.FLUID_SOURCE, List.of(ContextType.SERVER_WORLD, NetworkControllerContext.TYPE), graph);
+public class FluidTargetNode extends InterfaceNode {
+    public FluidTargetNode(Graph graph) {
+        super(ModNodeTypes.FLUID_TARGET, List.of(ContextType.SERVER_WORLD, NetworkControllerContext.TYPE), graph);
     }
 
     @Override
     public Connector<?>[] getOutputs() {
-        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredOutput("fluid", this) };
+        return new Connector[0];
     }
 
     @Override
     public Connector<?>[] getInputs() {
-        return new Connector[0];
+        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredInput("fluid", this) };
     }
 
     @Override
@@ -45,9 +42,10 @@ public class FluidSourceNode extends InterfaceNode {
 
         var storage = FluidStorage.SIDED.find(world, pos.offset(side), side.getOpposite());
         if (storage == null)
-            return Either.right(Text.translatable("node.advanced_networking.item_source.missing", interfaceId));
+            return Either.right(Text.translatable("node.advanced_networking.item_target.missing", interfaceId));
 
-        var stream = PathBundle.<Storage<FluidVariant>, FluidTransformer>begin(storage);
-        return Either.left(new DataValue<?>[] { ModDataTypes.FLUID_STREAM.makeValue(stream) });
+        var stream = inputs[0].getAs(ModDataTypes.FLUID_STREAM);
+        stream.end(storage, controller.controller().fluidEnvironment);
+        return Either.left(new DataValue<?>[0]);
     }
 }
