+++ src/main/java/io/github/mattidragon/advancednetworking/graph/node/item/filter/FilterItemsNode.java
@@ -1,10 +1,11 @@
-package io.github.mattidragon.advancednetworking.graph.node.fluid;
+package io.github.mattidragon.advancednetworking.graph.node.item.filter;
 
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.mojang.datafixers.util.Either;
 import io.github.mattidragon.advancednetworking.graph.ModDataTypes;
 import io.github.mattidragon.advancednetworking.graph.ModNodeTypes;
+import io.github.mattidragon.advancednetworking.graph.node.item.ItemTransformer;
 import io.github.mattidragon.nodeflow.graph.Connector;
 import io.github.mattidragon.nodeflow.graph.Graph;
 import io.github.mattidragon.nodeflow.graph.data.DataValue;
@@ -13,11 +14,11 @@ import io.github.mattidragon.nodeflow.ui.screen.EditorScreen;
 import io.github.mattidragon.nodeflow.ui.screen.NodeConfigScreen;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
-import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
+import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
 import net.minecraft.client.gui.widget.CyclingButtonWidget;
 import net.minecraft.client.gui.widget.TextFieldWidget;
 import net.minecraft.command.argument.NbtPathArgumentType;
-import net.minecraft.fluid.Fluid;
+import net.minecraft.item.Item;
 import net.minecraft.nbt.NbtCompound;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.RegistryKeys;
@@ -31,42 +32,42 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Predicate;
 
-public class FilterFluidNode extends Node {
-    private String fluidId = "";
+public class FilterItemsNode extends Node {
+    private String itemId = "";
     private String nbt = "";
-    private Mode mode = Mode.FLUID;
+    private Mode mode = Mode.ITEM;
     private boolean whitelist = true;
 
-    public FilterFluidNode(Graph graph) {
-        super(ModNodeTypes.FILTER_FLUID, List.of(), graph);
+    public FilterItemsNode(Graph graph) {
+        super(ModNodeTypes.FILTER_ITEMS, List.of(), graph);
     }
 
     @Override
     public Connector<?>[] getOutputs() {
-        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredOutput("out", this) };
+        return new Connector[] { ModDataTypes.ITEM_STREAM.makeRequiredOutput("out", this) };
     }
 
     @Override
     public Connector<?>[] getInputs() {
-        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredInput("in", this) };
+        return new Connector[] { ModDataTypes.ITEM_STREAM.makeRequiredInput("in", this) };
     }
 
     @Override
     public List<Text> validate() {
         var list = new ArrayList<Text>();
-        if (!fluidId.isBlank()) {
-            var id = Identifier.tryParse(fluidId.trim());
+        if (!itemId.isBlank()) {
+            var id = Identifier.tryParse(itemId.trim());
             if (id == null)
-                list.add(Text.translatable("node.advanced_networking.filter_fluid.invalid_id"));
-            else if (mode == Mode.FLUID && !Registries.FLUID.containsId(id))
-                list.add(Text.translatable("node.advanced_networking.filter_fluid.unknown_item", id));
+                list.add(Text.translatable("node.advanced_networking.filter_items.invalid_id"));
+            else if (mode == Mode.ITEM && !Registries.ITEM.containsId(id))
+                list.add(Text.translatable("node.advanced_networking.filter_items.unknown_item", id));
         }
 
         try {
             if (!nbt.isBlank())
                 NbtPathArgumentType.nbtPath().parse(new StringReader(nbt.trim()));
-        } catch (CommandSyntaxException e) {
-            list.add(Text.translatable("node.advanced_networking.filter_fluid.invalid_nbt_path", e.getMessage()));
+        } catch (CommandSyntaxException | StringIndexOutOfBoundsException e) {
+            list.add(Text.translatable("node.advanced_networking.filter_items.invalid_nbt_path", e.getMessage()));
         }
 
         list.addAll(super.validate());
@@ -76,8 +77,8 @@ public class FilterFluidNode extends Node {
 
     @Override
     protected Either<DataValue<?>[], Text> process(DataValue<?>[] inputs, ContextProvider context) {
-        var item = fluidId.isBlank() ? null : Registries.FLUID.get(new Identifier(fluidId));
-        var tag = fluidId.isBlank() ? null : TagKey.of(RegistryKeys.FLUID, new Identifier(fluidId));
+        var item = itemId.isBlank() ? null : Registries.ITEM.get(new Identifier(itemId));
+        var tag = itemId.isBlank() ? null : TagKey.of(RegistryKeys.ITEM, new Identifier(itemId));
         NbtPathArgumentType.NbtPath path;
         try {
             path = nbt.isBlank() ? null : NbtPathArgumentType.nbtPath().parse(new StringReader(nbt.trim()));
@@ -85,24 +86,24 @@ public class FilterFluidNode extends Node {
             throw new RuntimeException("Error while building nbt path not caught in validation", e);
         }
 
-        var stream = inputs[0].getAs(ModDataTypes.FLUID_STREAM);
-        stream.transform(new FluidTransformer.Filter(buildFilter(item, tag, path, whitelist)));
-        return Either.left(new DataValue<?>[]{ ModDataTypes.FLUID_STREAM.makeValue(stream) });
+        var stream = inputs[0].getAs(ModDataTypes.ITEM_STREAM);
+        stream.transform(new ItemTransformer.Filter(buildFilter(item, tag, path, whitelist)));
+        return Either.left(new DataValue<?>[]{ ModDataTypes.ITEM_STREAM.makeValue(stream) });
     }
 
     @NotNull
-    private Predicate<FluidVariant> buildFilter(Fluid fluid, TagKey<Fluid> tag, NbtPathArgumentType.NbtPath path, boolean whitelist) {
-        return (fluidVariant) -> {
-            if (mode == Mode.FLUID && fluid != null && fluidVariant.getFluid() != fluid)
+    private Predicate<ItemVariant> buildFilter(Item item, TagKey<Item> tag, NbtPathArgumentType.NbtPath path, boolean whitelist) {
+        return (itemVariant) -> {
+            if (mode == Mode.ITEM && item != null && itemVariant.getItem() != item)
                 return !whitelist;
 
             //noinspection deprecation
-            if (mode == Mode.TAG && tag != null && !fluidVariant.getFluid().getRegistryEntry().isIn(tag))
+            if (mode == Mode.TAG && tag != null && !itemVariant.getItem().getRegistryEntry().isIn(tag))
                 return !whitelist;
 
             if (path == null)
                 return whitelist;
-            if (path.count(fluidVariant.getNbt()) > 0)
+            if (path.count(itemVariant.getNbt()) > 0)
                 return whitelist;
             return !whitelist;
         };
@@ -111,7 +112,7 @@ public class FilterFluidNode extends Node {
     @Override
     public void readNbt(NbtCompound data) {
         super.readNbt(data);
-        fluidId = data.getString("fluidId");
+        itemId = data.getString("itemId");
         nbt = data.getString("nbt");
         mode = Mode.byOrdinal(data.getInt("mode"));
         whitelist = data.getBoolean("whitelist");
@@ -120,7 +121,7 @@ public class FilterFluidNode extends Node {
     @Override
     public void writeNbt(NbtCompound data) {
         super.writeNbt(data);
-        data.putString("fluidId", fluidId);
+        data.putString("itemId", itemId);
         data.putString("nbt", nbt);
         data.putInt("mode", mode.ordinal());
         data.putBoolean("whitelist", whitelist);
@@ -138,38 +139,38 @@ public class FilterFluidNode extends Node {
     }
 
     private enum Mode {
-        FLUID, TAG;
+        ITEM, TAG;
 
         private static Mode byOrdinal(int ordinal) {
-            return ordinal > 0 && ordinal < values().length ? values()[ordinal] : FLUID;
+            return ordinal > 0 && ordinal < values().length ? values()[ordinal] : ITEM;
         }
     }
 
     protected class ConfigScreen extends NodeConfigScreen {
         public ConfigScreen(EditorScreen parent) {
-            super(FilterFluidNode.this, parent);
+            super(FilterItemsNode.this, parent);
         }
 
         @Override
         protected void init() {
             var x = ((width - 200) / 2) - 50;
 
-            var whitelistButton = CyclingButtonWidget.onOffBuilder(Text.translatable("node.advanced_networking.filter_fluid.mode.whitelist"), Text.translatable("node.advanced_networking.filter_fluid.mode.blacklist"))
+            var whitelistButton = CyclingButtonWidget.onOffBuilder(Text.translatable("node.advanced_networking.filter_items.mode.whitelist"), Text.translatable("node.advanced_networking.filter_items.mode.blacklist"))
                     .initially(whitelist)
                     .omitKeyText()
                     .build(x, 70, 100, 20, Text.empty(), (button1, value) -> whitelist = value);
             addDrawableChild(whitelistButton);
 
-            var button = CyclingButtonWidget.<Mode>builder(mode -> mode == Mode.FLUID ? Text.translatable("node.advanced_networking.filter_fluid.mode.fluid") : Text.translatable("node.advanced_networking.filter_fluid.mode.tag"))
+            var button = CyclingButtonWidget.<Mode>builder(mode -> mode == Mode.ITEM ? Text.translatable("node.advanced_networking.filter_items.mode.item") : Text.translatable("node.advanced_networking.filter_items.mode.tag"))
                     .values(Mode.values())
                     .initially(mode)
-                    .build(x, 95, 100, 20, Text.translatable("node.advanced_networking.filter_fluid.mode"), (button1, value) -> mode = value);
+                    .build(x, 95, 100, 20, Text.translatable("node.advanced_networking.filter_items.mode"), (button1, value) -> mode = value);
             addDrawableChild(button);
 
             var idField = new TextFieldWidget(textRenderer, x, 120, 100, 20, Text.empty());
             idField.setPlaceholder(Text.literal("id").formatted(Formatting.GRAY));
-            idField.setText(fluidId);
-            idField.setChangedListener(newValue -> fluidId = newValue);
+            idField.setText(itemId);
+            idField.setChangedListener(newValue -> itemId = newValue);
             addDrawableChild(idField);
 
             var nbtField = new TextFieldWidget(textRenderer, x, 145, 100, 20, Text.empty());
