+++ src/main/java/io/github/mattidragon/advancednetworking/graph/node/item/storage/EmptyItemStreamNode.java
@@ -1,27 +1,28 @@
-package io.github.mattidragon.advancednetworking.graph.node.fluid;
+package io.github.mattidragon.advancednetworking.graph.node.item.storage;
 
 import com.mojang.datafixers.util.Either;
 import io.github.mattidragon.advancednetworking.graph.ModDataTypes;
 import io.github.mattidragon.advancednetworking.graph.ModNodeTypes;
+import io.github.mattidragon.advancednetworking.graph.node.item.ItemTransformer;
 import io.github.mattidragon.advancednetworking.graph.path.PathBundle;
 import io.github.mattidragon.nodeflow.graph.Connector;
 import io.github.mattidragon.nodeflow.graph.Graph;
 import io.github.mattidragon.nodeflow.graph.data.DataValue;
 import io.github.mattidragon.nodeflow.graph.node.Node;
-import net.fabricmc.fabric.api.transfer.v1.fluid.FluidVariant;
+import net.fabricmc.fabric.api.transfer.v1.item.ItemVariant;
 import net.fabricmc.fabric.api.transfer.v1.storage.Storage;
 import net.minecraft.text.Text;
 
 import java.util.List;
 
-public class EmptyFluidStreamNode extends Node {
-    public EmptyFluidStreamNode(Graph graph) {
-        super(ModNodeTypes.EMPTY_FLUID_STREAM, List.of(), graph);
+public class EmptyItemStreamNode extends Node {
+    public EmptyItemStreamNode(Graph graph) {
+        super(ModNodeTypes.EMPTY_ITEM_STREAM, List.of(), graph);
     }
 
     @Override
     public Connector<?>[] getOutputs() {
-        return new Connector[] { ModDataTypes.FLUID_STREAM.makeRequiredOutput("fluid", this) };
+        return new Connector[] { ModDataTypes.ITEM_STREAM.makeRequiredOutput("items", this) };
     }
 
     @Override
@@ -31,7 +32,7 @@ public class EmptyFluidStreamNode extends Node {
 
     @Override
     protected Either<DataValue<?>[], Text> process(DataValue<?>[] inputs, ContextProvider context) {
-        var stream = PathBundle.<Storage<FluidVariant>, FluidTransformer>begin(Storage.empty());
-        return Either.left(new DataValue<?>[] { ModDataTypes.FLUID_STREAM.makeValue(stream) });
+        var stream = PathBundle.<Storage<ItemVariant>, ItemTransformer>begin(Storage.empty());
+        return Either.left(new DataValue<?>[] { ModDataTypes.ITEM_STREAM.makeValue(stream) });
     }
 }
