+++ src/main/java/io/github/mattidragon/advancednetworking/graph/node/item/storage/ItemTargetNode.java
@@ -1,33 +1,32 @@
-package io.github.mattidragon.advancednetworking.graph.node.energy;
+package io.github.mattidragon.advancednetworking.graph.node.item.storage;
 
 import com.mojang.datafixers.util.Either;
 import io.github.mattidragon.advancednetworking.graph.ModDataTypes;
 import io.github.mattidragon.advancednetworking.graph.ModNodeTypes;
 import io.github.mattidragon.advancednetworking.graph.NetworkControllerContext;
 import io.github.mattidragon.advancednetworking.graph.node.InterfaceNode;
-import io.github.mattidragon.advancednetworking.graph.path.PathBundle;
 import io.github.mattidragon.nodeflow.graph.Connector;
 import io.github.mattidragon.nodeflow.graph.Graph;
 import io.github.mattidragon.nodeflow.graph.context.ContextType;
 import io.github.mattidragon.nodeflow.graph.data.DataValue;
+import net.fabricmc.fabric.api.transfer.v1.item.ItemStorage;
 import net.minecraft.text.Text;
-import team.reborn.energy.api.EnergyStorage;
 
 import java.util.List;
 
-public class EnergySourceNode extends InterfaceNode {
-    public EnergySourceNode(Graph graph) {
-        super(ModNodeTypes.ENERGY_SOURCE, List.of(ContextType.SERVER_WORLD, NetworkControllerContext.TYPE), graph);
+public class ItemTargetNode extends InterfaceNode {
+    public ItemTargetNode(Graph graph) {
+        super(ModNodeTypes.ITEM_TARGET, List.of(ContextType.SERVER_WORLD, NetworkControllerContext.TYPE), graph);
     }
 
     @Override
     public Connector<?>[] getOutputs() {
-        return new Connector[] { ModDataTypes.ENERGY_STREAM.makeRequiredOutput("energy", this) };
+        return new Connector[0];
     }
 
     @Override
     public Connector<?>[] getInputs() {
-        return new Connector[0];
+        return new Connector[] { ModDataTypes.ITEM_STREAM.makeRequiredInput("items", this) };
     }
 
     @Override
@@ -41,11 +40,12 @@ public class EnergySourceNode extends InterfaceNode {
         var pos = optionalPos.get().pos();
         var side = optionalPos.get().side();
 
-        var storage = EnergyStorage.SIDED.find(world, pos.offset(side), side.getOpposite());
+        var storage = ItemStorage.SIDED.find(world, pos.offset(side), side.getOpposite());
         if (storage == null)
-            return Either.right(Text.translatable("node.advanced_networking.energy_source.missing", interfaceId));
+            return Either.right(Text.translatable("node.advanced_networking.item_target.missing", interfaceId));
 
-        var bundle = PathBundle.<EnergyStorage, EnergyLimitTransformer>begin(storage);
-        return Either.left(new DataValue<?>[] { ModDataTypes.ENERGY_STREAM.makeValue(bundle) });
+        var stream = inputs[0].getAs(ModDataTypes.ITEM_STREAM);
+        stream.end(storage, controller.controller().itemEnvironment);
+        return Either.left(new DataValue<?>[0]);
     }
 }
